refactor: update src files structure
