+++ backend/src/app.module.ts
@@ -1,0 +1,7 @@
+import { Module } from '@nestjs/common';
+import { HelloWorldModule } from './hello-world/hello-world.module';
+
+@Module({
+  imports: [HelloWorldModule],
+})
+export class AppModule {}
