+++ backend/src/hello-world/hello-world.module.ts
@@ -1,0 +1,9 @@
+import { Module } from '@nestjs/common';
+import { HelloWorldController } from './hello-world.controller';
+import { HelloWorldService } from './hello-world.service';
+
+@Module({
+  controllers: [HelloWorldController],
+  providers: [HelloWorldService],
+})
+export class HelloWorldModule {}
