+++ backend/src/hello-world/hello-world.service.ts
@@ -1,0 +1,8 @@
+import { Injectable } from '@nestjs/common';
+
+@Injectable()
+export class HelloWorldService {
+  getHello(): string {
+    return '<h1> Hello world! </h1>';
+  }
+}
