--- backend/src/modules/app.module.ts
@@ -1,7 +1,0 @@
-import { Module } from '@nestjs/common';
-import { HelloWorldController } from './hello-world/hello-world.controller';
-
-@Module({
-  controllers: [HelloWorldController],
-})
-export class AppModule {}
